LVT_generated PET reader for LIS
This reader allows the user to use LVT-generated PET data as
forcing data in LIS (mainly for use with the geoWRSI model). The
dataset is currently on a global domain at 0.125x0.125 deg and
was generated using the Hobbins formulation for PET. This reader
will read in any type of PET as long as it was generated in LVT.
Future work for this reader will be to add a feature that either reads
in or allows the user to specify the data domain size and resolution to
add flexibility.