GDAS reader fix when downscaling
There were two issues identified with the GDAS reader that caused errors when downscaling
GDAS driver data.
The first issue occurs when GDAS data is being used to temporally downscale meteorological fields
(in our case, the issue was first identified when trying to downscale CHIRPS precip data). The issue
causes the GDAS reader to not 'refresh' the precipitation field at every timestep and only uses the final
precipitation field in a day to downscale the CHIRPS data.
To fix this issue, a secondary flag (gdas_struc%reset_flag) was added to force the GDAS reader to read in a new
precipitation field throughout the day in order to do the proper downscaling.

The second issue occurs when the LDT timestep is smaller than the GDAS data output. According to the GDAS reader
documentation in LDT, the reader should be reading in the forecast +03h data file to get the average meteorological field;
however, this was not the case.  The 'timenow' conditional in get_gdas.F90 was changed to .ge. instead of .gt. This allows
the reader to get the correct averaged field from the f+03h GDAS file, which is reflected in the logs and personal tests.

A third issue that did not have an effect on the reader was the conditional in create_gdasfilename.F90 where the 'F06flag'
variable was incorrecly being set to .true. when it should be .false.  This fix did not have an effect on the final reader because
the conditionals later in the code are written so that there is effectively another check and 'F06flag' is not used when
'is_analysis_hr' is true.